investments-microservice/requirements.txt
fastapi>=0.112.2
uvicorn>=0.30.6
pydantic>=2.8.2
psycopg2-binary>=2.9.9
python-dotenv
python-jose[cryptography]
